+++ FuzzyFilePath.py
@@ -11,7 +11,25 @@
             SHOULD BE:
             require("../../../../optimizer")|cursor|
 
-    @version 0.0.8
+    # bugs
+
+        - $module does not trigger completions
+        - wrong match:
+            FFP  --> trigger insert_path    component = require("./com"); ['component', (86, 95)]
+            FFP  <-- insert insert_path ['././$components', (90, 105)]
+
+    # errors
+
+        14/10/27
+
+            Traceback (most recent call last):
+              File "/Applications/Sublime Text.app/Contents/MacOS/sublime_plugin.py", line 374, in on_text_command
+                res = callback.on_text_command(v, name, args)
+              File "/Users/Gott/Dropbox/Applications/SublimeText/Packages/FuzzyFilePath/FuzzyFilePath.py", line 180, in on_text_command
+                Completion.before = re.sub(word_replaced + "$", "", path[0])
+            sre_constants.error: unbalanced parenthesis
+
+    @version 0.0.9
     @author Sascha Goldhofer <post@saschagoldhofer.de>
 """
 import sublime
@@ -19,220 +37,189 @@ import sublime_plugin
 import re
 import os
 
+import FuzzyFilePath.context as context
 from FuzzyFilePath.Cache.ProjectFiles import ProjectFiles
 from FuzzyFilePath.Query import Query
-
-DEBUG = False
-DISABLE_AUTOCOMPLETION = False
-DISABLE_KEYMAP_ACTIONS = False
-def verbose(*args):
-    if DEBUG is True:
-        print("FFP\t", *args)
-
-Completion = {
-
-    "active": False,
-    "before": None,
-    "after": None,
-    "onInsert": []
-}
+from FuzzyFilePath.common.verbose import verbose
+from FuzzyFilePath.common.config import config
 
 query = Query()
 project_files = None
 
+class Completion:
+    active = False
+    before = None
+    after = None
+    onInsert = []
+
+    def reset():
+        Completion.before = None
+        Completion.replaceOnInsert = []
+
+    def get_final_path(path):
+        if Completion.before is not None:
+            Completion.before = re.escape(Completion.before)
+            path = re.sub("^" + Completion.before, "", path)
+        # hack reverse
+        path = re.sub(config["ESCAPE_DOLLAR"], "$", path)
+        for replace in Completion.replaceOnInsert:
+            path = re.sub(replace[0], replace[1], path)
+        return path
+
 
 def plugin_loaded():
     """load settings"""
-    settings = sublime.load_settings("FuzzyFilePath.sublime-settings")
+    settings = sublime.load_settings(config["FFP_SETTINGS_FILE"])
     settings.add_on_change("extensionsToSuggest", update_settings)
     update_settings()
 
 
 def update_settings():
     """restart projectFiles with new plugin and project settings"""
-    global project_files, DISABLE_AUTOCOMPLETION, DISABLE_KEYMAP_ACTIONS
+    global project_files, config
 
     exclude_folders = []
     project_folders = sublime.active_window().project_data().get("folders", [])
-    settings = sublime.load_settings("FuzzyFilePath.sublime-settings")
+    settings = sublime.load_settings(config["FFP_SETTINGS_FILE"])
     query.scopes = settings.get("scopes", [])
     query.auto_trigger = (settings.get("auto_trigger", True))
-    DISABLE_AUTOCOMPLETION = settings.get("disable_autocompletions", False);
-    DISABLE_KEYMAP_ACTIONS = settings.get("disable_keymap_actions", False);
-
-    # build exclude folders
-    for folder in project_folders:
-        base = folder.get("path")
-        exclude = folder.get("folder_exclude_patterns", [])
-        for f in exclude:
-            exclude_folders.append(os.path.join(base, f))
-    # or use default settings
-    if (len(exclude_folders) == 0):
-        exclude_folders = settings.get("excludeFolders", ["node_modules"])
-
+    exclude_folders = settings.get("exclude_folders", ["node_modules"])
     project_files = ProjectFiles(settings.get("extensionsToSuggest", ["js"]), exclude_folders)
 
-
-def get_path_at_cursor(view):
-    word = get_word_at_cursor(view)
-    line = get_line_at_cursor(view)
-    path = get_path(line[0], word[0])
-    path_region = sublime.Region(word[1].a, word[1].b)
-    path_region.b = word[1].b
-    path_region.a = word[1].a - (len(path) - len(word[0]))
-    verbose("path_at_cursor", path, "word:", word, "line", line)
-    return [path, path_region]
-
-
-# tested
-def get_path(line, word):
-    #! returns first match
-    if word is None or word is "":
-        return word
-
-    needle = re.escape(word)
-    full_words = line.split(" ")
-    for full_word in full_words:
-        if word in line:
-            path = extract_path_from(full_word, needle)
-            if not path is None:
-                return path
-
-    return word
-
-
-#! fails if needle occurs also before path (line)
-def extract_path_from(word, needle):
-    result = re.search('([^\"\'\s]*)' + needle + '([^\"\'\s]*)', word)
-    if (result is not None):
-        return result.group(0)
-    return None
-
-
-def get_line_at_cursor(view):
-    selection = view.sel()[0]
-    position = selection.begin()
-    region = view.line(position)
-    line = view.substr(region)
-    verbose("line at cursor", line)
-    return [line, region]
-
-
-# tested
-def get_word_at_cursor(view):
-    selection = view.sel()[0]
-    position = selection.begin()
-    region = view.word(position)
-    word = view.substr(region)
-    # validate
-    valid = not re.sub("[\"\'\s\(\)]*", "", word).strip() == ""
-    if not valid:
-        verbose("invalid word", word)
-        return ["", sublime.Region(position, position)]
-    # single line only
-    if "\n" in word:
-        return ["", sublime.Region(position, position)]
-    # strip quotes
-    if len(word) > 0:
-        if word[0] is '"':
-            word = word[1:]
-            region.a += 1
-
-        if word[-1:] is '"':
-            word = word[1:]
-            region.a += 1
-    # cleanup in case an empty string is encounterd
-    if word.find("''") != -1 or word.find('""') != -1 or word.isspace():
-        word = ""
-        region = sublime.Region(position, position)
-
-    return [word, region]
-
-
-class ReplaceRegionCommand(sublime_plugin.TextCommand):
-    # helper: replaces range with string
-    def run(self, edit, a, b, string):
-        if DISABLE_KEYMAP_ACTIONS is True:
-            return False
-        self.view.replace(edit, sublime.Region(a, b), string)
+    config["DISABLE_KEYMAP_ACTIONS"] = settings.get("disable_keymap_actions", config["DISABLE_KEYMAP_ACTIONS"]);
+    config["DISABLE_AUTOCOMPLETION"] = settings.get("disable_autocompletions", config["DISABLE_AUTOCOMPLETION"]);
 
 
 class InsertPathCommand(sublime_plugin.TextCommand):
+
     # triggers autocomplete
     def run(self, edit, type="default", replace_on_insert=[]):
-        if DISABLE_KEYMAP_ACTIONS is True:
+        if config["DISABLE_KEYMAP_ACTIONS"] is True:
             return False
 
         query.relative = type
         if len(replace_on_insert) > 0:
             verbose("insert path", "override replace", replace_on_insert)
             query.override_replace_on_insert(replace_on_insert)
-        self.view.run_command('auto_complete')
+
+        self.view.run_command('auto_complete', "insert")
 
 
 class FuzzyFilePath(sublime_plugin.EventListener):
 
+    """
+        track and validate: on_post_insert_completion
+    """
+    track_insert = {
+        "active": False,
+        "start_line": "",
+        "start_path": "",
+        "end_line": "",
+        "end_path": ""
+    }
+
+    def start_tracking(self, view, command_name=None):
+        self.track_insert["active"] = True
+        self.track_insert["start_line"] = context.get_line_at_cursor(view)[0]
+        self.track_insert["end_line"] = None
+        self.track_insert["start_path"] = context.get_path_at_cursor(view)
+        verbose("--> trigger", command_name, self.track_insert["start_line"], self.track_insert["start_path"])
+
+        path = context.get_path_at_cursor(view)
+        word_replaced = re.split("[./]", path[0]).pop()
+        if (path is not word_replaced):
+            Completion.before = re.sub(word_replaced + "$", "", path[0])
+
+    def finish_tracking(self, view, command_name=None):
+        self.track_insert["active"] = False
+        self.track_insert["end_line"] = context.get_line_at_cursor(view)[0]
+        self.track_insert["end_path"] = context.get_path_at_cursor(view)
+        verbose("<-- insert", command_name, self.track_insert["end_path"])
+
+    def abort_tracking(self):
+        self.track_insert["active"] = False
+
     def on_text_command(self, view, command_name, args):
-        if command_name == "commit_completion":
-            path = get_path_at_cursor(view)
-            word_replaced = re.split("[./]", path[0]).pop()
-            if (path is not word_replaced):
-                Completion["before"] = re.sub(word_replaced + "$", "", path[0])
-
+        # check if a completion may be inserted
+        if command_name in config["TRIGGER_ACTION"] or command_name in config["INSERT_ACTION"]:
+            self.start_tracking(view, command_name)
         elif command_name == "hide_auto_complete":
-            Completion["active"] = False
-
+            Completion.active = False
+            self.abort_tracking()
 
     def on_post_text_command(self, view, command_name, args):
-        if (command_name == "commit_completion" and Completion["active"]):
-            Completion["active"] = False
-            # replace current path (fragments) with selected path
-            # i.e. ../../../file -> ../file
-            # if Completion["before"] is not None:
-            path = get_path_at_cursor(view)
-            final_path = re.sub("^" + Completion["before"], "", path[0])
-            # hack reverse
-            # final_path = final_path.replace("_D011AR_", "$")
-            final_path = re.sub("_D011AR_", "$", final_path)
-            # modify result
-            for replace in Completion["replaceOnInsert"]:
-                final_path = re.sub(replace[0], replace[1], final_path)
-            # cleanup path
-            view.run_command("replace_region", { "a": path[1].a, "b": path[1].b, "string": final_path })
-            #reset
-            Completion["before"] = None
-            Completion["replaceOnInsert"] = []
+        # check if a completion is inserted
+        current_line = context.get_line_at_cursor(view)[0]
+        insert = command_name in config["TRIGGER_ACTION"] and self.track_insert["start_line"] != current_line
+        insert = insert or command_name in config["INSERT_ACTION"]
 
+        if insert is True:
+            self.finish_tracking(view, command_name)
+            self.on_post_insert_completion(view, command_name)
 
-    def on_post_save_async(self, view):
-        if project_files is not None:
-            for folder in sublime.active_window().folders():
-                if folder in view.file_name():
-                    project_files.update(folder, view.file_name())
-
-
+    """
+        query filepath completion
+    """
     def on_query_completions(self, view, prefix, locations):
-        if (DISABLE_AUTOCOMPLETION is True):
+        # check if a completion may be inserted
+        if self.track_insert["active"] is False:
+            self.start_tracking(view)
+
+        if (config["DISABLE_AUTOCOMPLETION"] is True):
             return None
 
         if query.valid is False:
             return False
 
-        needle = get_path_at_cursor(view)[0]
+        needle = context.get_path_at_cursor(view)[0]
         current_scope = view.scope_name(locations[0])
 
         if query.build(current_scope, needle, query.relative) is False:
             return None
 
-        view.run_command('_enter_insert_mode') # vintageous
-        Completion["active"] = True
         completions = project_files.search_completions(query.needle, query.project_folder, query.extensions, query.relative, query.extension)
-        verbose("query needle:", needle, "relative:", query.relative)
-        verbose("query completions:", completions)
-        Completion["replaceOnInsert"] = query.replace_on_insert
+
+        if len(completions) > 0:
+            verbose("completions", len(completions), "found for", query.needle)
+            Completion.active = True
+            Completion.replaceOnInsert = query.replace_on_insert
+            # vintageous
+            view.run_command('_enter_insert_mode')
+        else:
+            verbose("completions", "no completions for", query.needle)
+            Completion.active = False
+
         query.reset()
         return completions
 
+    """
+        post filepath completion
+    """
+    def on_post_insert_completion(self, view, command_name):
+        """ Sanitize inserted path by
+            - replacing temporary variables (~$)
+            - replacing query partials, like "../<inserted path>"
+        """
+        if Completion.active is False:
+            return None
+
+        Completion.active = False
+        path = context.get_path_at_cursor(view)
+        # remove path query completely
+        final_path = Completion.get_final_path(path[0])
+        # replace current query with final path
+        view.run_command("ffp_replace_region", { "a": path[1].a, "b": path[1].b, "string": final_path })
+
+        Completion.reset()
+
+    """
+        update cached files
+    """
+    def on_post_save_async(self, view):
+        if project_files is not None:
+            for folder in sublime.active_window().folders():
+                if folder in view.file_name():
+                    project_files.update(folder, view.file_name())
 
     def on_activated(self, view):
         file_name = view.file_name()