+++ common/config.py
@@ -1,0 +1,11 @@
+config = {
+
+	"DEBUG": False,
+	"FFP_SETTINGS_FILE": "FuzzyFilePath.sublime-settings",
+	"ESCAPE_DOLLAR": '\$',
+	"TRIGGER_ACTION": ["auto_complete", "insert_path"],
+	"INSERT_ACTION": ["commit_completion", "insert_best_completion"],
+	# dynamic
+	"DISABLE_AUTOCOMPLETION": False,
+	"DISABLE_KEYMAP_ACTIONS": False
+}